dispatcher/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-jose[cryptography]
